// apps/busy/migrations/snapshots/44a0c79b03b7cc0cf8f2467fbe10154fbf1c9d060e29c67964e2325949e464f0/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Connection: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly created_at: CodecTypes['pg/timestamptz@1']['output'];
      readonly updated_at: CodecTypes['pg/timestamptz@1']['output'];
      readonly type: CodecTypes['pg/text@1']['output'];
      readonly accessToken: CodecTypes['pg/text@1']['output'] | null;
      readonly refreshToken: CodecTypes['pg/text@1']['output'] | null;
      readonly expiresAt: CodecTypes['pg/timestamptz@1']['output'] | null;
      readonly deviceId: CodecTypes['pg/text@1']['output'];
    };
    readonly Device: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly created_at: CodecTypes['pg/timestamptz@1']['output'];
      readonly updated_at: CodecTypes['pg/timestamptz@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Connection: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly created_at: CodecTypes['pg/timestamptz@1']['input'];
      readonly updated_at: CodecTypes['pg/timestamptz@1']['input'];
      readonly type: CodecTypes['pg/text@1']['input'];
      readonly accessToken: CodecTypes['pg/text@1']['input'] | null;
      readonly refreshToken: CodecTypes['pg/text@1']['input'] | null;
      readonly expiresAt: CodecTypes['pg/timestamptz@1']['input'] | null;
      readonly deviceId: CodecTypes['pg/text@1']['input'];
    };
    readonly Device: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly created_at: CodecTypes['pg/timestamptz@1']['input'];
      readonly updated_at: CodecTypes['pg/timestamptz@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly 'ayrock.busy.connection': {
      readonly accessToken: CodecTypes['pg/text@1']['output'] | null;
      readonly created_at: CodecTypes['pg/timestamptz@1']['output'];
      readonly deviceId: CodecTypes['pg/text@1']['output'];
      readonly expiresAt: CodecTypes['pg/timestamptz@1']['output'] | null;
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly refreshToken: CodecTypes['pg/text@1']['output'] | null;
      readonly type: CodecTypes['pg/text@1']['output'];
      readonly updated_at: CodecTypes['pg/timestamptz@1']['output'];
    };
    readonly 'ayrock.busy.device': {
      readonly created_at: CodecTypes['pg/timestamptz@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'] | null;
      readonly updated_at: CodecTypes['pg/timestamptz@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly 'ayrock.busy.connection': {
      readonly accessToken: CodecTypes['pg/text@1']['input'] | null;
      readonly created_at: CodecTypes['pg/timestamptz@1']['input'];
      readonly deviceId: CodecTypes['pg/text@1']['input'];
      readonly expiresAt: CodecTypes['pg/timestamptz@1']['input'] | null;
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly refreshToken: CodecTypes['pg/text@1']['input'] | null;
      readonly type: CodecTypes['pg/text@1']['input'];
      readonly updated_at: CodecTypes['pg/timestamptz@1']['input'];
    };
    readonly 'ayrock.busy.device': {
      readonly created_at: CodecTypes['pg/timestamptz@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'] | null;
      readonly updated_at: CodecTypes['pg/timestamptz@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;